cloture du lot3

+++ src/AppBundle/Controller/Rest/InvitationController.php
@@ -4,12 +4,17 @@ namespace AppBundle\Controller\Rest;
 
 use AppBundle\Document\Invitation;
 use AppBundle\Service\InvitationService;
-use AppBundle\Service\SearchGuardService;
 use ATS\CoreBundle\Controller\Rest\RestControllerTrait;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use AppBundle\Service\ApplicationService;
+use Symfony\Component\HttpKernel\Exception\HttpException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
+use AppBundle\Document\AccessLevel;
 
 class InvitationController extends Controller
 {
@@ -47,21 +52,35 @@ class InvitationController extends Controller
     }
 
     /**
-     * @Route("/new", methods="POST")
+     * @Route("/{envId}/new", methods="POST")
      *
      * @param Request $request
      * @param InvitationService $invitationService
-     *
+     * @param ApplicationService $applicationService
+     * @param string $envId
      * @return Response
      */
-    public function newInvitationAction(Request $request, InvitationService $invitationService)
+    public function newInvitationAction(
+        Request $request, 
+        InvitationService $invitationService,
+        ApplicationService $applicationService,
+        $envId)
     {
+        $user = $this->getUser();
         $data = $request->getContent();
-        $successful = $invitationService->newInvitation($data, $this->getUser());
+        $app = json_decode($data, true);
 
-        return $this->renderResponse($successful);
+        if($applicationService->userHasPermission(
+            $app["application"]["id"], 
+            $user, 
+            $envId, 
+            array(AccessLevel::ADMIN))){
+            $successful = $invitationService->newInvitation($data, $this->getUser());
+            return $this->renderResponse($successful);
+        } else {
+            return $this->exception(['message' => "You dont have rights permissions"], 400);
+        }
     }
-
     /**
      * @Route("/{id}/update", methods="PUT")
      *
@@ -79,19 +98,35 @@ class InvitationController extends Controller
     }
 
     /**
-     * @Route("/{id}/delete", methods="DELETE")
+     * @Route("/{id}/delete", methods="POST")
      *
      * @param Request $request
      * @param InvitationService $invitationService
+     * @param ApplicationService $applicationService
      * @param string $id
      *
      * @return Response
      */
-    public function deleteInvitationAction(Request $request, InvitationService $invitationService, $id)
+    public function deleteInvitationAction(
+        Request $request, 
+        InvitationService $invitationService,
+        ApplicationService $applicationService,
+        $id)
     {
-        $invitationService->deleteInvitation($id);
+        $data = $request->getContent();
+        $_data = \json_decode($data, true);
+        $user = $this->getUser();
+        if($applicationService->userHasPermission(
+            $_data["appId"], 
+            $user, 
+            $_data["envId"],
+            array(AccessLevel::ADMIN))){
+                $invitationService->deleteInvitation($id);
+                return $this->renderResponse(null);
+        } else {
+            return $this->exception(['message' => "You dont have rights permissions"], 400);
+        }
 
-        return $this->renderResponse(null);
     }
 
     /**
@@ -99,7 +134,6 @@ class InvitationController extends Controller
      *
      * @param Request $request
      * @param InvitationService $invitationService
-     * @param SearchGuardService $sgService
      * @param string $id
      *
      * @return Response
@@ -107,13 +141,10 @@ class InvitationController extends Controller
     public function acceptInvitationAction(
         Request $request,
         InvitationService $invitationService,
-        SearchGuardService $sgService,
         $id
     ) {
         $user = json_decode($request->getContent());
         $invitationService->acceptInvitation($id, $user->userId);
-        $sgService->updateSearchGuardConfig();
-
         return $this->renderResponse(null);
     }
 }
